Merging from cefc1fd2d3fc1a0c336babedbf00dcbece928aea:
Change label service suggestion to single line

The suggestion only inserts a single triple, and it is uncommon to add
more, so I don’t think the service needs to take up three lines in the
query by default. The single-line version also doesn’t need to be
adjusted for indentation.

Change-Id: I6c3fd3be5dee6a169c9d347c67e3e7a71c3e12fb