Merging from 35663e8731371f73a05cb513e227a5b6d8cdc05b:
Introduce query templates

The template definition is a JSON string embedded as a SPARQL comment.
A template defines a textual representation for a query and variables
that can be changed by the user.
Each variable can optionally provide a SPARQL query for elements to be
proposed to the user.

See example:
http://tinyurl.com/hfhdp88

Change-Id: I737dc03811e57016ad787b8769f4e89e4a09b57a