Merging from 244c505e9f692fe6831d6ac03d88f794eaab524f:
Only hide “scroll to top” in embed.html

The fix for T178222 in commit 07aa3535ba (change I53b82b1b5f) hides all
Leaflet controls in the top right corner, which is too heavy-handed – we
want to hide the “scroll to top” button, but not the layers control.
Instead, add a custom class to the “scroll to top” button and only hide
that button specifically, leaving the layers control in place.

Bug: T193930
Change-Id: I563b7cf7cbbaa783e5062c38370ffeaf1275131c