Merging from 871c1a3efabee66d17c53549d62fe90b9070b851:
Fixes problem with WDQS HTML table export for cells without value

When downloading a result set in WDQS as HTML table, the cells that don't contain any values are not created as table cells in the HTML table by using <td> elements. This causes the cells to not properly be aligned when columns containing optional values are on the left or in the middle.

Bug: T207257
Change-Id: I0823189a26447005c1c1d8a1d6325b1e211ca9fe