Support wdqs and wcqs from the same deployment
The wdqs microsite deployment is being generalized into query_service
deployment to enable support for wcqs. Does not yet install anything
for wcqs, this is preparation to make that possible.

Support is provided by creating a per-site directory that will host
assets specific to that deployment such as the favicon and
custom-config.json. Removes the prior assets in the root of the repo,
and .gitignore's so that the CI build process does not re-add the
generic imagery from the main repo.

This patch must be deployed after the webserver configuration is put
in place to fallback from the old location to the new location when
not found. Out of caution this patch does not remove
custom-config.json. Once deployed and we verify assets are correctly
served from the new locations a followup can remove the json config.

Bug: T280247
Depends-On: I609ee42c2b163f852686503b748e5292405f4cde
Change-Id: I87a52a46589b4e1739e5c7370b7bf0a0908db630