Merging from db18c9ad6d8a993fc93f859a5ac56434989a017e
Set html font-size to 1rem

Bootstrap 3 sets the html font-size to 10px; it’s not really used by
anything in the query UI directly (the font-size gets overridden
starting at the body), but is relevant for the rem unit, which Leaflet
has started to use since version 1.8.0. Set it to 1em to restore the
user agent font size (unset or initial wouldn’t work in IE11).

Bug: T311139
Co-Authored-By: Michael Große <michael.grosse@wikimedia.de>
Change-Id: Ia024f63e3800d6e5d3ba244a81f01136611ae43e