Merging from 319bd6f39d3482a25fe01bdde766df7f1cc92f40
Adjust query builder button and its popup

Order of buttons has been changed so that the query
builder button is the last one.
Popup image was removed and style adjusted, so it
is smaller and less intrusive.
Minimal width for the dropdown window set to em so it remains
relativ (as opposed to e.g. fonts, that should stay absolute
and are therefore set in px)

Bug: T296135
Change-Id: I9b9ad8a92cf90f0cda5c406297a383711f720fed