Add configuration for experimental split graph UI microsites
As part of an experiment to split the graph, we're creating
the various UI so that we can easily run test queries manually
on those endpoints.

The naming is making it clear that this is experimental, proper
naming will be required once in production.

Note that the icons and logo are copied from WDQS, we will need
to change those eventually.

Bug: T354658
Change-Id: I393ad4de6b43f4836ba1aec64048fe530db1df3d