Merging from 77e858ba31cee5fb0b621cc44ff30852420fca0a
Partially revert package upgrades

This partially reverts change I71c0edee86 (commit fbfbc9f8db) –
apparently something in here broke some of the Dimple-based charts.

I produced this by checking out the commit prior to that change and then
restoring the package.json changes bit by bit, testing the query GUI
build in between – I was expecting to find the package upgrade that
broke things, but instead I ended up with the new versions of all the
packages mentioned in package.json and everything still worked. So it
seems that just one of the indirect dependencies in package-lock.json
caused the problem. Maybe we can figure out which one it is later, but
for now let’s fix the bug, I think.

Bug: T375058
Change-Id: Iecb34e5d3b2d1d4ae11eb3bddc6a8a7dafd6f56a